The server now treats a disconnect as a resign

+++ protocol/src/main/java/com/nathanholmberg/chess/protocol/messages/game/server/ClockUpdateMessage.java
@@ -6,7 +6,7 @@ public class ClockUpdateMessage extends Message {
     private final long whiteTime;
     private final long blackTime;
 
-    public ClockUpdateMessage(int whiteTime, int blackTime) {
+    public ClockUpdateMessage(long whiteTime, long blackTime) {
         this.whiteTime = whiteTime;
         this.blackTime = blackTime;
     }
