+++ protocol/src/main/java/com/nathanholmberg/chess/protocol/messages/game/server/GameStateMessage.java
@@ -3,7 +3,7 @@ package com.nathanholmberg.chess.protocol.messages.game.server;
 public class GameStateMessage extends ClockUpdateMessage {
     private final String pgn;
 
-    public GameStateMessage(String pgn, int whiteTime, int blackTime) {
+    public GameStateMessage(String pgn, long whiteTime, long blackTime) {
         super(whiteTime, blackTime);
         this.pgn = pgn;
     }
